incorporating working files into project
