+++ Makefile
@@ -6,7 +6,7 @@
 #    By: kwurster <kwurster@student.42berlin.de>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/04/22 15:55:27 by lmeubrin          #+#    #+#              #
-#    Updated: 2025/05/05 13:42:37 by lmeubrin         ###   ########.fr        #
+#    Updated: 2025/05/05 19:36:37 by lmeubrin         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -92,7 +92,7 @@ SRCS := $(addprefix $(SRCS_DIR)/,\
 		scene_malloc.c scene_destroy.c interval.c preprocess.c preprocess2.c \
 		print_scene.c print_scene_utils_0.c print_scene_utils_1.c) \
 		$(addprefix parse/, input_check.c parse_scene.c set_scene.c set_objs.c \
-		set_utils.c) \
+		set_utils.c set_bumpmap.c get_texture.c set_textures.c) \
 		)
 
 OBJS := $(SRCS:($SRCS_DIR)%.c=$(OBJ_DIR)/%.o)
