+++ include/parse.h
@@ -6,7 +6,7 @@
 /*   By: lmeubrin <lmeubrin@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/19 11:29:00 by lmeubrin          #+#    #+#             */
-/*   Updated: 2025/05/05 16:26:47 by lmeubrin         ###   ########.fr       */
+/*   Updated: 2025/05/05 19:37:52 by lmeubrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,3 +60,6 @@ void	print_cones(const t_cone *cones, size_t cone_count);
 void	*rperror_get_next_line(char *line, char *errormessage);
 int		set_bumpmap(char *obj_line, t_sphere *sphere);
 int		open_bumpmap_file(char *line);
+
+int	get_val(int fd);
+t_bumpmap	*allocate_bumpmap(int fd);
