+++ srcs/parse/get_texture.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   get_bumpmap.c                                      :+:      :+:    :+:   */
+/*   get_texture.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lmeubrin <lmeubrin@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/05/05 16:09:04 by lmeubrin          #+#    #+#             */
-/*   Updated: 2025/05/05 18:13:07 by lmeubrin         ###   ########.fr       */
+/*   Created: 2025/05/05 19:35:47 by lmeubrin          #+#    #+#             */
+/*   Updated: 2025/05/05 19:38:00 by lmeubrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ static	char	*parse_file_head(int fd)
 
 // starts parsing the file to get file id (only accepting P3 for .ppm)
 // gets width and height to allocate struct
-static t_bumpmap	*allocate_bumpmap(int fd)
+t_bumpmap	*allocate_bumpmap(int fd)
 {
 	char		*line;
 	t_bumpmap	*bump;
@@ -80,7 +80,7 @@ static t_bumpmap	*allocate_bumpmap(int fd)
 	return (bump);
 }
 
-static	int	get_val(int fd)
+int	get_val(int fd)
 {
 	char	*line;
 	char	*endptr;
@@ -100,48 +100,3 @@ static	int	get_val(int fd)
 	return (num);
 }
 
-static int	fill_bump_greyscale(t_bumpmap *bump, int fd)
-{
-	int		r;
-	int		g;
-	int		b;
-	int		i;
-	int		max_pixels;
-
-	i = 0;
-	max_pixels = bump->height * bump->width;
-	while (i <= max_pixels)
-	{
-		r = get_val(fd);
-		if (r == -1)
-			break ;
-		g = get_val(fd);
-		if (g == -1)
-			return (0);
-		b = get_val(fd);
-		if (b == -1)
-			return (0);
-		bump->elevation[i] = ((float)r + (float)g + (float)b) / (3.0f * 255.0f);
-		++i;
-	}
-	// get_next_line(-1);
-	return (1);
-}
-
-int	set_bumpmap(char *obj_line, t_sphere *sphere)
-{
-	int			fd;
-	char		*line;
-	t_bumpmap	*bump;
-
-	fd = open_bumpmap_file(obj_line);
-	if (fd == -1)
-		return (0);
-	bump = allocate_bumpmap(fd);
-	if (!bump)
-		return (0);
-	if(!fill_bump_greyscale(bump, fd))
-		return (0);
-	sphere->bumpmap = bump;
-	return (1);
-}
