+++ srcs/util/scene_destroy.c
@@ -6,11 +6,19 @@
 /*   By: lmeubrin <lmeubrin@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/28 12:30:26 by lmeubrin          #+#    #+#             */
-/*   Updated: 2025/05/05 11:27:55 by lmeubrin         ###   ########.fr       */
+/*   Updated: 2025/05/05 19:34:03 by lmeubrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/miniRT_types.h"
+
+void	bumpmap_destroy(t_bumpmap *bumpmap)
+{
+	if (!bumpmap)
+		return ;
+	free(bumpmap->elevation);
+	free(bumpmap);
+}
 
 static void	objects_destroy_inner(t_objects *objects)
 {
@@ -20,7 +28,12 @@ static void	objects_destroy_inner(t_objects *objects)
 		return ;
 	i = 0;
 	while (i < objects->sphere_count)
-		free(objects->spheres[i++].lights);
+	{
+		free(objects->spheres[i].lights);
+		if (objects->spheres->texture == BUMP)
+			bumpmap_destroy(objects->spheres[i].bumpmap);
+		i++;
+	}
 	i = 0;
 	while (i < objects->cylinder_count)
 		free(objects->cylinders[i++].lights);
@@ -35,22 +48,14 @@ void	objects_destroy(t_objects *objects, bool free_inner)
 		return ;
 	if (free_inner)
 		objects_destroy_inner(objects);
-	free(objects->planes);
-	objects->plane_count = 0;
 	free(objects->spheres);
 	objects->sphere_count = 0;
+	free(objects->planes);
+	objects->plane_count = 0;
 	free(objects->cylinders);
 	objects->cylinder_count = 0;
 	free(objects->cones);
 	objects->cone_count = 0;
-}
-
-void	bumpmap_destroy(t_bumpmap *bumpmap)
-{
-	if (!bumpmap)
-		return ;
-	free(bumpmap->elevation);
-	free(bumpmap);
 }
 
 /// Free memory allocated for a scene
